BOL-122 refactoring: user rank item RV 관련 파일들 패키지 이동
